requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0,<3.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pyjwt>=2.8.0
cryptography>=41.0.0
pyyaml>=6.0
prometheus-client>=0.18.0
zstandard>=0.22.0

